Add bcond for OBS-specific quirks
The version substitution system is not able to fully subst
the current Version field due to the inline use of macros, so you end up with like:

257-123-gabcd257.1

instead of:

257-123-gabcd

I.E., the hard-coded 257.1 gets appended to the OBS-specified
version.

If it was simply hardcoded as 257.1 it would work, but the inline
macros throw it off.